equalsで違う型を比較してるところを修正、== でオブジェクトを比較しているところを修正
git-svn-id: http://aipo.googlecode.com/svn/trunk@2403 b7d4e767-949e-c8c6-5735-6c11b8a2f74f

+++ portlets/report/src/main/java/com/aimluck/eip/report/ReportSelectData.java
@@ -335,7 +335,8 @@ public class ReportSelectData extends
       // 自身の報告書かを設定する
       Integer login_user_id =
         Integer.valueOf((int) login_user.getUserId().getValue());
-      rd.setIsSelfReport(record.getUserId() == login_user_id);
+      rd.setIsSelfReport(record.getUserId().intValue() == login_user_id
+        .intValue());
       return rd;
     } catch (Exception ex) {
       logger.error("Exception", ex);
